requirements.txt
asgiref==3.5.2
black==22.3.0
boto3==1.24.14
botocore==1.27.14
click==8.1.3
coverage==6.4
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==3.2
django-allauth==0.50.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.1
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
redgreenunittest==0.1.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
sqlparse==0.4.2
stripe==3.4.0
